requirements.txt
td-watson
python-dateutil
click >=7.0,<8.0
simplejson
colorama
jira
pyyaml
pyxdg
